src/Industrial_control/app_snap7/requirements.txt
python-snap7
httpx>=0.25.2